Also block all legacy modules from being imported on the command line
