+++ api/src/opentrons/__init__.py
@@ -1,4 +1,3 @@
-import importlib
 import json
 import os
 import sys
@@ -12,6 +11,7 @@ from opentrons.hardware_control import API, ThreadManager
 from opentrons.config import (feature_flags as ff, name,
                               robot_configs, IS_ROBOT, ROBOT_FIRMWARE_DIR)
 from opentrons.util import logging_config
+from opentrons.protocols.types import ApiDeprecationError, APIVersion
 
 version = sys.version_info[0:2]
 if version < (3, 7):
@@ -33,7 +33,7 @@ from opentrons import config  # noqa(E402)
 LEGACY_MODULES = [
     'robot', 'reset', 'instruments', 'containers', 'labware', 'modules']
 
-__all__ = ['version', 'HERE', 'LEGACY_MODULES', 'config']
+__all__ = ['version', 'HERE', 'config']
 
 
 def __getattr__(attrname):
@@ -46,15 +46,18 @@ def __getattr__(attrname):
     this function won't be invoked.
     """
     if attrname in LEGACY_MODULES:
-        legacy_api = importlib.import_module(
-            '.'.join([__name__,
-                      'legacy_api',
-                      'api']))
-        for mod in LEGACY_MODULES:
-            setattr(sys.modules[__name__], attrname,
-                    getattr(legacy_api, attrname))
-        return getattr(sys.modules[__name__], attrname)
+        raise ApiDeprecationError(APIVersion(1, 0))
     raise AttributeError(attrname)
+    # if attrname in LEGACY_MODULES:
+    #     legacy_api = importlib.import_module(
+    #         '.'.join([__name__,
+    #                     'legacy_api',
+    #                     'api']))
+    #     for mod in LEGACY_MODULES:
+    #         setattr(sys.modules[__name__], attrname,
+    #                 getattr(legacy_api, attrname))
+    #     return getattr(sys.modules[__name__], attrname)
+    # raise AttributeError(attrname)
 
 
 def __dir__():
