--- api/src/opentrons/commands/__init__.py
@@ -1,1 +1,0 @@
-from .commands import *  # noqa
