+++ api/src/opentrons/protocol_api/protocol_context.py
@@ -4,11 +4,12 @@ import logging
 from typing import (Dict, Iterator, List,
                     Optional, Set, Tuple, Union, TYPE_CHECKING)
 
-from opentrons import types, commands as cmds
+from opentrons import types
 from opentrons.hardware_control import (SynchronousAdapter, modules,
                                         API, ExecutionManager)
 from opentrons.config import feature_flags as fflags
-from opentrons.commands import CommandPublisher
+from opentrons.commands import protocol_commands as cmds, types as cmd_types
+from opentrons.commands.publisher import CommandPublisher, publish
 from opentrons.protocols.api_support.constants import SHORT_TRASH_DECK, \
     STANDARD_DECK
 from opentrons.protocols.types import APIVersion, Protocol
@@ -223,7 +224,7 @@ class ProtocolContext(CommandPublisher):
                 self._commands.append(text.format(**payload))
 
         self._unsubscribe_commands = self.broker.subscribe(
-            cmds.types.COMMAND, on_command)
+            cmd_types.COMMAND, on_command)
 
     @contextlib.contextmanager
     def temp_connect(self, hardware: API):
@@ -594,7 +595,7 @@ class ProtocolContext(CommandPublisher):
                 in self._instruments.items()
                 if instr}
 
-    @cmds.publish.both(command=cmds.pause)
+    @publish.both(command=cmds.pause)
     @requires_version(2, 0)
     def pause(self, msg=None):
         """ Pause execution of the protocol until resume is called.
@@ -607,13 +608,13 @@ class ProtocolContext(CommandPublisher):
         """
         self._hw_manager.hardware.pause()
 
-    @cmds.publish.both(command=cmds.resume)
+    @publish.both(command=cmds.resume)
     @requires_version(2, 0)
     def resume(self):
         """ Resume a previously-paused protocol """
         self._hw_manager.hardware.resume()
 
-    @cmds.publish.both(command=cmds.comment)
+    @publish.both(command=cmds.comment)
     @requires_version(2, 0)
     def comment(self, msg):
         """
@@ -626,7 +627,7 @@ class ProtocolContext(CommandPublisher):
         """
         pass
 
-    @cmds.publish.both(command=cmds.delay)
+    @publish.both(command=cmds.delay)
     @requires_version(2, 0)
     def delay(self, seconds=0, minutes=0, msg=None):
         """ Delay protocol execution for a specific amount of time.
