+++ api/src/opentrons/protocols/api_support/definitions.py
@@ -5,3 +5,6 @@ MAX_SUPPORTED_VERSION = types.APIVersion(2, 8)
 
 V2_MODULE_DEF_VERSION = types.APIVersion(2, 3)
 #: The first API version in which we prefer the v2 module defs
+
+MINIMUM_SUPPORTED_VERSION = types.APIVersion(2, 0)
+#: The minimum supported protocol API version in this release
