+++ api/tests/opentrons/conftest.py
@@ -23,8 +23,6 @@ import zipfile
 
 import pytest
 
-from opentrons.legacy_api.containers import load
-from opentrons.legacy_api.instruments.pipette import Pipette
 from opentrons.api.routers import MainRouter
 from opentrons.api import models
 from opentrons.data_storage import database_migration
@@ -33,10 +31,6 @@ from opentrons import hardware_control as hc
 from opentrons.hardware_control import API, ThreadManager, ThreadedAsyncLock
 from opentrons.protocol_api import ProtocolContext
 from opentrons.types import Mount, Location, Point
-from opentrons import (robot as rb,
-                       instruments as ins,
-                       containers as cns,
-                       modules as mods)
 
 
 Session = namedtuple(
@@ -46,37 +40,6 @@ Session = namedtuple(
 Protocol = namedtuple(
     'Protocol',
     ['text', 'filename', 'filelike'])
-
-
-@pytest.fixture
-def singletons(virtual_smoothie_env):
-    rb.reset()
-    yield {'robot': rb,
-           'instruments': ins,
-           'labware': cns,
-           'modules': mods}
-    rb.disconnect()
-    rb.reset()
-
-
-@pytest.fixture
-def robot(singletons):
-    yield singletons['robot']
-
-
-@pytest.fixture
-def instruments(singletons):
-    yield singletons['instruments']
-
-
-@pytest.fixture
-def labware(singletons):
-    yield singletons['labware']
-
-
-@pytest.fixture
-def modules(singletons):
-    yield singletons['modules']
 
 
 @pytest.fixture(autouse=True)
@@ -340,21 +303,6 @@ async def wait_until(matcher, notifications, timeout=1, loop=None):
             return result
 
 
-def build_v1_model(r, lw_name):
-    plate = load(r, lw_name or '96-flat', '1')
-    tiprack = load(r, 'opentrons-tiprack-300ul', '2')
-    pipette = Pipette(r,
-                      ul_per_mm=18.5, max_volume=300, mount='right',
-                      tip_racks=[tiprack])
-    instrument = models.Instrument(pipette)
-    container = models.Container(plate)
-    return namedtuple('model', 'robot instrument container')(
-        robot=r,
-        instrument=instrument,
-        container=container,
-    )
-
-
 def build_v2_model(h, lw_name, loop):
     ctx = ProtocolContext(loop=loop, hardware=h)
 
@@ -375,14 +323,11 @@ def build_v2_model(h, lw_name, loop):
     )
 
 
-@pytest.fixture(params=[build_v1_model, build_v2_model])
-def model(request, robot, hardware, loop):
+@pytest.fixture(params=[build_v2_model])
+def model(request, hardware, loop):
     # Use with pytest.mark.parametrize(’labware’, [some-labware-name])
     # to have a different labware loaded as .container. If not passed,
     # defaults to the version-appropriate way to do 96 flat
-    if request.node.get_closest_marker('api1_only')\
-       and request.param != build_v1_model:
-        pytest.skip('only works with a robot')
     if request.node.get_closest_marker('api2_only')\
        and request.param != build_v2_model:
         pytest.skip('only works with hardware controller')
@@ -393,28 +338,7 @@ def model(request, robot, hardware, loop):
 
     builder = request.param
 
-    if builder == build_v1_model:
-        return builder(robot, lw_name)
-    else:
-        return builder(hardware, lw_name, loop)
-
-
-@pytest.fixture
-def model_with_trough(robot):
-    from opentrons.legacy_api.containers import load
-    from opentrons.legacy_api.instruments.pipette import Pipette
-
-    pipette = Pipette(robot, ul_per_mm=18.5, max_volume=300, mount='right')
-    plate = load(robot, 'trough-12row', '1')
-
-    instrument = models.Instrument(pipette)
-    container = models.Container(plate)
-
-    return namedtuple('model', 'robot instrument container')(
-            robot=robot,
-            instrument=instrument,
-            container=container
-        )
+    return builder(hardware, lw_name, loop)
 
 
 @pytest.fixture
