+++ api/tests/opentrons/drivers/test_driver.py
@@ -1,16 +1,11 @@
 from copy import deepcopy
-from threading import Thread
 from unittest.mock import Mock
 import pytest
-from time import sleep
 
-from numpy import isclose
-
-from opentrons.trackers import pose_tracker
 from tests.opentrons.conftest import fuzzy_assert
 from opentrons.config.robot_configs import (
     DEFAULT_GANTRY_STEPS_PER_MM, DEFAULT_PIPETTE_CONFIGS)
-from opentrons.drivers import serial_communication, utils, types
+from opentrons.drivers import serial_communication
 from opentrons.drivers.smoothie_drivers import driver_3_0
 
 
@@ -431,86 +426,6 @@ def test_functional(smoothie):
     assert smoothie.position == smoothie.homed_position
 
 
-@pytest.mark.api1_only
-def test_set_pick_up_current(model, monkeypatch):
-    driver = model.robot._driver
-
-    set_current = driver._save_current
-    current_log = []
-
-    def set_current_mock(target, axes_active=True):
-        nonlocal current_log
-        current_log.append(target)
-        set_current(target, axes_active)
-
-    monkeypatch.setattr(driver, '_save_current', set_current_mock)
-    driver.update_homed_flags({ax: True for ax in 'XYZABC'})
-
-    rack = model.robot.add_container('tiprack-200ul', '10')
-    pipette = model.instrument._instrument
-    pipette.set_pick_up_current(0.42)
-    pipette.pick_up_tip(rack[0], presses=1)
-
-    # Instrument in `model` is configured to right mount, which is the A axis
-    # on the Smoothie (see `Robot._actuators`)
-    expected = [
-        {'C': 0.5},
-        {'C': 0.05},
-        {'A': 0.8},
-        {'A': 0.1},
-        {'X': 1.25, 'Y': 1.25},
-        {'X': 0.3, 'Y': 0.3},
-        {'A': 0.8},
-        {'A': 0.42},
-        {'A': 0.8},
-        {'A': 0.1}
-    ]
-    assert current_log == expected
-
-
-@pytest.mark.xfail
-@pytest.mark.api1_only
-def test_drop_tip_current(model, monkeypatch):
-    # TODO: All of these API 1 tests either need to be removed or moved to
-    # a different test file. The ones using the model fixture rely on
-    # some ugly things created in RPC. Ideally, all of these tests should
-    # be testing methods in the smoothie directly.
-    driver = model.driver
-
-    old_save_current = driver._save_current
-    current_log = []
-
-    def mock_save_current(settings, axes_active=True):
-        nonlocal current_log
-        if 'C' in settings:
-            current_log.append(settings)
-        old_save_current(settings, axes_active)
-
-    monkeypatch.setattr(driver, '_save_current', mock_save_current)
-
-    rack = model.robot.add_container('tiprack-200ul', '10')
-    pipette = model.instrument._instrument
-    pipette._plunger_current = 0.123
-    pipette._drop_tip_current = 0.456
-    pipette.drop_tip(rack[0])
-
-    # Instrument in `model` is configured to right mount, which is the A axis
-    # on the Smoothie (see `Robot._actuators`)
-    expected = [
-        {'C': 0.123},   # move to 'bottom' position
-        {'C': 0.05},    # dwell
-        {'C': 0.456},   # move to 'drop_tip' position
-        {'C': 0.05},    # dwell
-        {'C': 0.123},   # fast-home move upwards
-        {'C': 0.05},    # dwell
-        {'C': 0.123},   # fast-home home command
-        {'C': 0.05},    # dwell
-        {'C': 0.123},   # move back to 'bottom' position
-        {'C': 0.05}     # dwell
-    ]
-    assert current_log == expected
-
-
 def test_parse_pipette_data():
     msg = 'TestsRule!!'
     mount = 'L'
@@ -708,438 +623,3 @@ def test_clear_limit_switch(smoothie, monkeypatch):
         'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005',
         'M400',
     ]
-
-
-@pytest.mark.api1_only
-def test_pause_resume(model):
-    """
-    This test has to use an ugly work-around with the `simulating` member of
-    the driver. When issuing movement commands in test, `simulating` should be
-    True, but when testing whether `pause` actually pauses and `resume`
-    resumes, `simulating` must be False.
-    """
-
-    pipette = model.instrument._instrument
-    robot = model.robot
-
-    robot.home()
-    homed_coords = pose_tracker.absolute(robot.poses, pipette)
-
-    robot._driver.simulating = False
-    robot.pause()
-    robot._driver.simulating = True
-
-    def _move_head():
-        robot.poses = pipette._move(robot.poses, x=100, y=0, z=0)
-
-    thread = Thread(target=_move_head)
-    thread.start()
-    sleep(0.5)
-
-    # Check against home coordinates before calling resume to ensure that robot
-    # doesn't move while paused
-    coords = pose_tracker.absolute(robot.poses, pipette)
-    assert isclose(coords, homed_coords).all()
-
-    robot._driver.simulating = False
-    robot.resume()
-    robot._driver.simulating = True
-    thread.join()
-
-    coords = pose_tracker.absolute(robot.poses, pipette)
-    expected_coords = (100, 0, 0)
-    assert isclose(coords, expected_coords).all()
-
-
-def test_speed_change(robot, instruments, monkeypatch):
-    ulmm = {
-        "aspirate": [[100, 0, 0.5]],
-        "dispense": [[100, 0, 0.5]]
-    }
-    pipette = instruments.Pipette(mount='right', ul_per_mm=ulmm)
-    robot._driver.simulating = False
-
-    command_log = []
-
-    def write_with_log(command, ack, connection, timeout, tag=None):
-        if 'G0F' in command:
-            command_log.append(command.strip())
-        elif 'M114' in command:
-            return 'ok MCS: X:0.00 Y:0.00 Z:0.00 A:0.00 B:0.00 C:0.00'
-        return driver_3_0.SMOOTHIE_ACK
-
-    monkeypatch.setattr(serial_communication, 'write_and_return',
-                        write_with_log)
-
-    pipette.tip_attached = True
-    pipette.max_volume = 100
-    pipette._working_volume = 100
-    pipette.set_speed(aspirate=20, dispense=40)
-    pipette.aspirate(10)
-    pipette.dispense(10)
-    expected = [
-        ['G0F1200'],  # pipette's default aspirate speed in mm/min
-        ['G0F24000'],
-        ['G0F2400'],  # pipette's default dispense speed in mm/min
-        ['G0F24000'],
-    ]
-    fuzzy_assert(result=command_log, expected=expected)
-
-
-def test_max_speed_change(robot, smoothie, monkeypatch):
-    smoothie.simulating = False
-    robot._driver = smoothie
-
-    from opentrons.drivers import serial_communication
-    from opentrons.drivers.smoothie_drivers import driver_3_0
-    command_log = []
-
-    def write_with_log(command, ack, connection, timeout, tag=None):
-        if 'M203.1' in command or 'G0F' in command:
-            command_log.append(command.strip())
-        return driver_3_0.SMOOTHIE_ACK
-
-    monkeypatch.setattr(serial_communication, 'write_and_return',
-                        write_with_log)
-
-    robot.head_speed(555)
-    robot.head_speed(x=1, y=2, z=3, a=4, b=5, c=6)
-    robot.head_speed(123, x=7)
-    robot._driver.push_speed()
-    robot._driver.set_speed(321)
-    robot._driver.pop_speed()
-    expected = [
-        ['G0F{}'.format(555 * 60)],
-        ['M203.1 A4 B5 C6 X1 Y2 Z3'],
-        ['M203.1 X7'],
-        ['G0F{}'.format(123 * 60)],
-        ['G0F{}'.format(321 * 60)],
-        ['G0F{}'.format(123 * 60)],
-    ]
-    fuzzy_assert(result=command_log, expected=expected)
-
-
-@pytest.mark.api1_only
-def test_pause_in_protocol(model):
-    model.robot._driver.simulating = True
-
-    model.robot.pause()
-
-    assert model.robot._driver.run_flag.is_set()
-
-
-def test_send_command_with_retry(robot, smoothie, monkeypatch):
-    smoothie.simulating = False
-    robot._driver = smoothie
-
-    count = 0
-
-    def _no_response(command, ack, connection, timeout, tag=None):
-        nonlocal count
-        count += 1
-        if count < 3:
-            raise serial_communication.SerialNoResponse('No response')
-        else:
-            return 'ok'
-
-    monkeypatch.setattr(serial_communication, 'write_and_return', _no_response)
-
-    # force `write_and_return` to raise exception just once
-    count = 0
-    res = robot._driver._send_command('test')
-    assert res == 'ok'
-
-    # force `write_and_return` to raise exception twice
-    count = -1
-    with pytest.raises(serial_communication.SerialNoResponse):
-        robot._driver._send_command('test')
-
-
-def test_unstick_axes(robot, smoothie):
-    import types
-
-    smoothie.simulating = False
-    robot._driver = smoothie
-
-    def update_position_mock(self, default=None):
-        if default is None:
-            default = self._position
-
-        updated_position = self._position.copy()
-        updated_position.update(**default)
-
-    robot._driver.update_position = types.MethodType(
-        update_position_mock, robot._driver)
-
-    current_log = []
-
-    def send_command_mock(self, command, timeout=12000.0, ack_timeout=5.0):
-        nonlocal current_log
-        current_log.append(command)
-        if 'M119' in command:
-            smoothie_switch_res = 'X_max:0 Y_max:0 Z_max:0 A_max:0 B_max:0 C_max:0'  # NOQA
-            smoothie_switch_res += ' _pins '
-            smoothie_switch_res += '(XL)2.01:0 (YL)2.01:0 (ZL)2.01:0 '
-            smoothie_switch_res += '(AL)2.01:0 (BL)2.01:0 (CL)2.01:0 Probe: 0\r\n'   # NOQA
-            return smoothie_switch_res
-
-    robot._driver._send_command = types.MethodType(
-        send_command_mock, robot._driver)
-
-    robot._driver.unstick_axes('BC')
-
-    expected = [
-        'M203.1 B1 C1',  # slow them down
-        'M119',  # get the switch status
-        # move
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 G0B-1C-1',
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005',  # set plunger current
-        'M203.1 A125 B40 C40 X600 Y400 Z125'  # return to normal speed
-    ]
-
-    assert current_log == expected
-
-    current_log = []
-    robot._driver.unstick_axes('XYZA')
-
-    expected = [
-        'M203.1 A1 X1 Y1 Z1',  # slow them down
-        'M119',  # get the switch status
-        'M907 A0.8 B0.05 C0.05 X1.25 Y1.25 Z0.8 G4P0.005 G0A-1X-1Y-1Z-1',  # noqa(E501)
-        'M203.1 A125 B40 C40 X600 Y400 Z125'  # return to normal speed
-    ]
-
-    assert current_log == expected
-
-    def send_command_mock(self, command, timeout=12000.0, ack_timeout=5.0):
-        nonlocal current_log
-        current_log.append(command)
-        if 'M119' in command:
-            smoothie_switch_res = 'X_max:0 Y_max:0 Z_max:0 A_max:0 B_max:0 C_max:1'  # NOQA
-            smoothie_switch_res += ' _pins '
-            smoothie_switch_res += '(XL)2.01:0 (YL)2.01:0 (ZL)2.01:0 '
-            smoothie_switch_res += '(AL)2.01:0 (BL)2.01:0 (CL)2.01:0 Probe: 0\r\n'   # NOQA
-            return smoothie_switch_res
-
-    robot._driver._send_command = types.MethodType(
-        send_command_mock, robot._driver)
-
-    current_log = []
-    robot._driver.unstick_axes('BC')
-
-    expected = [
-        'M203.1 B1 C1',  # set max-speeds
-        'M119',  # get switch status
-        # MOVE B
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 G0B-2',
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005',  # low current B
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 G28.2C',  # HOME C
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005',  # low current C
-        'M203.1 A125 B40 C40 X600 Y400 Z125'  # reset max-speeds
-    ]
-    assert current_log == expected
-
-    def send_command_mock(self, command, timeout=12000.0, ack_timeout=5.0):
-        nonlocal current_log
-        current_log.append(command)
-        if 'M119' in command:
-            smoothie_switch_res = 'X_max:0 Y_max:0 Z_max:0 A_max:0 B_max:1 C_max:1'  # NOQA
-            smoothie_switch_res += ' _pins '
-            smoothie_switch_res += '(XL)2.01:0 (YL)2.01:0 (ZL)2.01:0 '
-            smoothie_switch_res += '(AL)2.01:0 (BL)2.01:0 (CL)2.01:0 Probe: 0\r\n'   # NOQA
-            return smoothie_switch_res
-
-    robot._driver._send_command = types.MethodType(
-        send_command_mock, robot._driver)
-
-    current_log = []
-    robot._driver.unstick_axes('BC')
-
-    expected = [
-        'M203.1 B1 C1',  # set max-speeds
-        'M119',  # get switch status
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 G28.2BC',  # HOME BC
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005',  # low current BC
-        'M203.1 A125 B40 C40 X600 Y400 Z125'  # reset max-speeds
-    ]
-    assert current_log == expected
-
-
-def test_alarm_unhandled(smoothie, robot, monkeypatch):
-    smoothie.simulating = False
-    robot._driver = smoothie
-    killmsg = 'ALARM: Kill button pressed - reset or M999 to continue\r\n'
-
-    def fake_write_and_return(cmdstr, ack, conn, timeout=None, tag=None):
-
-        return cmdstr + killmsg
-
-    monkeypatch.setattr(serial_communication, 'write_and_return',
-                        fake_write_and_return)
-    assert serial_communication.write_and_return is fake_write_and_return
-    robot._driver.move({'X': 0})
-
-    robot._driver._is_hard_halting.set()
-
-    with pytest.raises(driver_3_0.SmoothieAlarm):
-        robot._driver.move({'X': 25})
-
-    assert not robot._driver._is_hard_halting.is_set()
-
-
-def test_move_splitting(smoothie, robot, monkeypatch):
-    smoothie.simulating = False
-    command_log = []
-
-    time_mock = Mock()
-    monkeypatch.setattr(utils.time, 'monotonic', time_mock)
-    time_mock.return_value = 0
-
-    def send_command_logger(command, timeout=12000.0, ack_timeout=5.0):
-        nonlocal command_log
-        command_log.append(command)
-
-    monkeypatch.setattr(smoothie, '_send_command', send_command_logger)
-    smoothie.update_steps_per_mm({'B': 3200, 'C': 3200})
-    command_log.clear()
-
-    time_mock.return_value = 10
-    smoothie.move({'X': 100})
-    # no backlash, no move splitting, nice and easy
-    assert command_log\
-        == ['M907 A0.1 B0.05 C0.05 X1.25 Y0.3 Z0.1 G4P0.005 G0X100']
-
-    command_log.clear()
-
-    # move splitting but for a different axis - ignored
-    smoothie.configure_splits_for({'B': types.MoveSplit(
-        split_distance=50, split_current=1.5, split_speed=0.5, after_time=0,
-        fullstep=True)})
-    time_mock.return_value = 20
-    smoothie.move({'C': 10})
-    assert command_log\
-        == ['M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 G0C10.3 G0C10',
-            'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005']
-
-    command_log.clear()
-
-    # move splits that are longer than the move get eaten both in -
-    time_mock.return_value = 40
-    smoothie.configure_splits_for(
-        {'B': types.MoveSplit(
-            split_distance=30, split_current=1.5,
-            split_speed=0.5, after_time=0, fullstep=False)})
-    smoothie._position['B'] = 100
-    smoothie.move({'B': 75})
-    assert command_log\
-        == ['G0F30 M907 A0.1 B1.5 C0.05 X0.3 Y0.3 Z0.1 G4P0.005',
-            'G0B75',
-            'G0F24000 M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 G0B75',
-            'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005']
-
-    command_log.clear()
-    # and in +
-    time_mock.return_value = 50
-    smoothie.move({'B': 100})
-    assert command_log\
-        == ['G0F30 M907 A0.1 B1.5 C0.05 X0.3 Y0.3 Z0.1 G4P0.005',
-            'G0B100.3',
-            'G0F24000 M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 '
-            'G0B100.3 G0B100',
-            'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005']
-
-    # if backlash is involved, it's added on top
-    # prep by moving to 0
-    time_mock.return_value = 60
-    smoothie.move({'C': 0})
-    smoothie.configure_splits_for({'C': types.MoveSplit(
-        split_distance=1, split_current=2.0, split_speed=1, after_time=0,
-        fullstep=True)})
-    command_log.clear()
-    smoothie.move({'C': 20})
-    assert command_log\
-        == ['M55 M92 C100.0 G4P0.01 '
-            'G0F60 M907 A0.1 B0.05 C2.0 X0.3 Y0.3 Z0.1 G4P0.005',
-            'G0C1',
-            'M54 M92 C3200 G4P0.01',
-            'G0F24000 M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 '
-            'G0C20.3 G0C20',
-            'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005']  # noqa(E501)
-
-    # if backlash is involved, the backlash target should be the limit
-    # for the split move
-    smoothie.move({'C': 15})
-    smoothie.configure_splits_for({'C': types.MoveSplit(
-        split_distance=10, split_current=2.0, split_speed=1, after_time=0,
-        fullstep=True)})
-    command_log.clear()
-    time_mock.return_value = 70
-    smoothie.move({'C': 20})
-    # note that the backlash/target move has a 0.05A current on C even though
-    # it is active because that is the robot config default active plunger
-    # current. when the driver is used with the rest of the robot or hardware
-    # control stack it uses the higher currents
-    assert command_log\
-        == ['M55 M92 C100.0 G4P0.01 '
-            'G0F60 M907 A0.1 B0.05 C2.0 X0.3 Y0.3 Z0.1 G4P0.005',
-            'G0C20.3',
-            'M54 M92 C3200 G4P0.01',
-            'G0F24000 M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 '
-            'G0C20.3 G0C20',
-            'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005']  # noqa(E501)
-    smoothie.configure_splits_for(
-        {'B': types.MoveSplit(
-            split_distance=50, split_current=1.5,
-            split_speed=0.5, after_time=10,
-            fullstep=True)})
-    # timing: if the axis has moved recently (since we're changing the
-    # time mock) it shouldn't split. first move to reset the last moved at
-    smoothie.move({'B': 0})
-    command_log.clear()
-    # this move therefore should not split
-    smoothie.move({'B': 100})
-    assert command_log[0:1] == [
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 G0B100.3 G0B100']
-    command_log.clear()
-    # nor should this move
-    time_mock.return_value = 79
-    smoothie.move({'B': 1})
-    assert command_log[0:1] == [
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 G0B1']
-    command_log.clear()
-    # now that we advance time, we split
-    time_mock.return_value = 89.01
-    command_log.clear()
-    smoothie.move({'B': 100})
-    assert command_log == [
-        'M53 M92 B100.0 G4P0.01 G0F30 '
-        'M907 A0.1 B1.5 C0.05 X0.3 Y0.3 Z0.1 G4P0.005',
-        'G0B51',
-        'M52 M92 B3200 G4P0.01',
-        'G0F24000 M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005 '
-        'G0B100.3 G0B100',
-        'M907 A0.1 B0.05 C0.05 X0.3 Y0.3 Z0.1 G4P0.005']
-
-
-def test_per_move_speed(smoothie, robot, monkeypatch):
-    smoothie.simulating = False
-    command_log = []
-
-    def send_command_logger(command, timeout=12000.0, ack_timeout=5.0):
-        nonlocal command_log
-        command_log.append(command)
-
-    monkeypatch.setattr(smoothie, '_send_command', send_command_logger)
-
-    # no speed argument: use combined speed
-    smoothie.move({'X': 100})
-
-    assert command_log[0]\
-        == 'M907 A0.1 B0.05 C0.05 X1.25 Y0.3 Z0.1 G4P0.005 G0X100'
-
-    command_log.clear()
-
-    # specify speed: both set and reset
-    smoothie.move({'Y': 100}, speed=100)
-    assert command_log[0]\
-        == 'G0F6000 M907 A0.1 B0.05 C0.05 X0.3 Y1.25 Z0.1 G4P0.005 G0Y100 G0F24000'  # noqa(E501)
