--- api/tests/opentrons/labware/data/calibrations.json
@@ -1,1 +1,0 @@
-{}
