--- api/tests/opentrons/labware/data/invalid_json.json
@@ -1,1 +1,0 @@
-{asdfsad}
