+++ robot-server/tests/service/legacy/rpc/test_server.py
@@ -492,68 +492,3 @@ def test_notifications(session, root):
             'data': 'Finishing'
         }
     ]
-
-
-@pytest.mark.api1_only
-@pytest.mark.parametrize('root', [ConcurrentCallTester()])
-def test_concurrent_call(session, root):
-    session.socket.receive_json()  # Skip init
-
-    session.call(
-        id=id(root),
-        name='start',
-        args=[]
-    )
-    session.socket.receive_json()  # Skip ack
-
-    # Get the notification that the function is starting
-    r = session.socket.receive_json()
-    assert r == {'$': {'type': rpc.NOTIFICATION_MESSAGE},
-                 'data': 'starting'}
-
-    # Call function to awaken the thread
-    session.call(
-        id=id(root),
-        name='resume',
-        args=[]
-    )
-
-    # Ack for call
-    r = session.socket.receive_json()
-    assert r == {'$': {'token': session.token,
-                       'type': rpc.CALL_ACK_MESSAGE}}
-
-    # Result of awaken call
-    r = session.socket.receive_json()
-    assert r == {'$': {'status': 'success', 'token': session.token,
-                       'type': rpc.CALL_RESULT_MESSAGE},
-                 'data': 'Resumed'}
-
-    # Notification of being awake
-    r = session.socket.receive_json()
-    assert r == {'$': {'type': rpc.NOTIFICATION_MESSAGE},
-                 'data': 'awake'}
-
-    # Awaken again
-    session.call(
-        id=id(root),
-        name='resume',
-        args=[]
-    )
-
-    # Ack for call
-    r = session.socket.receive_json()
-    assert r == {'$': {'token': session.token,
-                       'type': rpc.CALL_ACK_MESSAGE}}
-
-    # Result of awaken
-    r = session.socket.receive_json()
-    assert r == {'$': {'status': 'success', 'token': session.token,
-                       'type': rpc.CALL_RESULT_MESSAGE},
-                 'data': 'Resumed'}
-
-    # Result of start function
-    r = session.socket.receive_json()
-    assert r == {'$': {'status': 'success', 'token': session.token,
-                       'type': rpc.CALL_RESULT_MESSAGE},
-                 'data': 'Done!'}
